+++ tests/utils/debounce.test.js
@@ -5,45 +5,41 @@ describe('debounce', () => {
     /*
      * Jest.useFakeTimers() enables fake timers.
      * This is replacing the original implementation of setTimeout() and other timer functions.
-     * Timers can be restored to their normal behavior with jest.useRealTimers().
+     * Timers can be restored to their normal behavior with vi.useRealTimers().
      */
-    jest.useFakeTimers()
+    vi.useFakeTimers()
   })
-  it('doesnt call the function if it has a waiting time', () => {
-    const callback = jest.fn()
+  it('does not call the function if the debounce is reset within the cooldown time', () => {
+    const callback = vi.fn()
     const debounced = debounce(callback, 500)
-    debounced(callback)
+
+    debounced()
+    vi.advanceTimersByTime(400)
+
+    debounced() // reset
+
+    vi.advanceTimersByTime(499) // still in cooldown (1ms left)
     expect(callback).not.toBeCalled()
 
-    jest.advanceTimersByTime(100)
-    debounced();
-    expect(callback).not.toBeCalled()
-
-    jest.advanceTimersByTime(499)
-    expect(callback).not.toBeCalled()
-
-    jest.advanceTimersByTime(300)
+    vi.advanceTimersByTime(1) // cooldown finished
     expect(callback).toBeCalledTimes(1)
   })
 
-  it('fires the callback, when immediate is true, no matter of the waiting time', () => {
-    const callback = jest.fn();
+  it('fires immediately and ignores calls within the cooldown time when immediate is true', () => {
+    const callback = vi.fn()
     const debounced = debounce(callback, 500, true)
-    expect(callback).not.toBeCalled()
+
     debounced()
-    expect(callback).toBeCalledTimes(1)
+    expect(callback).toBeCalledTimes(1) // immediate execution
 
-    jest.advanceTimersByTime(100)
     debounced()
+    vi.advanceTimersByTime(100) // still in cooldown (400ms left)
     expect(callback).toBeCalledTimes(1)
 
-    jest.advanceTimersByTime(499)
+    vi.advanceTimersByTime(399) // just before cooldown ends (1ms left)
     expect(callback).toBeCalledTimes(1)
 
-    jest.advanceTimersByTime(1)
-    expect(callback).toBeCalledTimes(1)
-
-    jest.advanceTimersByTime(700)
+    vi.advanceTimersByTime(1) // cooldown finished
     debounced()
     expect(callback).toBeCalledTimes(2)
   })